bot/requirements.txt
anyio==4.12.1
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
nest-asyncio==1.6.0
python-dotenv==1.0.0
python-telegram-bot==21.0.1
typing_extensions==4.15.0
